requirements.txt
fastapi==0.95.2
fastapi-jwt-auth==0.5.0

uvicorn==0.22.0

sqlalchemy==1.4.22
aiomysql==0.1.1

pydantic
pydantic-settings

passlib==1.7.4
python-jose==3.3.0
pytz==2023.3
werkzeug==2.0.3
cryptography
python-multipart